requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-pptx==1.0.2
python-docx==0.8.11
beautifulsoup4==4.12.2
lxml==4.9.3
aiofiles==23.2.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cors==1.0.1
httpx==0.25.1
Pillow==10.1.0